feat: 🎸 SearchCondition create

+++ Back_end/jmark/src/main/java/com/common/jmark/service/CompanyService.java
@@ -1,21 +1,26 @@
 package com.common.jmark.service;
 
 import com.common.jmark.common.config.security.util.JwtUtil;
+import com.common.jmark.common.exception.NotAuthException;
 import com.common.jmark.common.exception.NotFoundException;
 import com.common.jmark.common.exception.NotMatchException;
+import com.common.jmark.domain.entity.Apply;
 import com.common.jmark.domain.entity.Company;
 import com.common.jmark.domain.entity.JobOpening;
 import com.common.jmark.domain.entity.category.Gugun;
 import com.common.jmark.domain.entity.category.JobChildCategory;
+import com.common.jmark.domain.repository.ApplyRepository;
 import com.common.jmark.domain.repository.CompanyRepository;
 import com.common.jmark.domain.repository.JobOpeningRepository;
 import com.common.jmark.domain.repository.category.GugunRepository;
 import com.common.jmark.domain.repository.category.JobChildCategoryRepository;
+import com.common.jmark.dto.ApplyDto;
 import com.common.jmark.dto.CompanyDto;
 import com.common.jmark.dto.CompanyLoginRequest;
 import com.common.jmark.dto.JobOpeningDto;
 import com.common.jmark.dto.category.GugunResponse;
 import com.common.jmark.dto.category.JobChildCategoryResponse;
+import com.common.jmark.dto.user.UserResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -24,6 +29,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import static com.common.jmark.common.exception.NotAuthException.COMPANY_NOT_MATCH;
 import static com.common.jmark.common.exception.NotFoundException.USER_NOT_FOUND;
 import static com.common.jmark.common.exception.NotMatchException.PASSWORD_NOT_MATCH;
 
@@ -35,6 +41,7 @@ public class CompanyService {
     private final CompanyRepository companyRepository;
     private final GugunRepository gugunRepository;
     private final JobChildCategoryRepository jobChildCategoryRepository;
+    private final ApplyRepository applyRepository;
     private final JwtUtil jwtUtil;
 
     //회사 회원가입
@@ -73,54 +80,42 @@ public class CompanyService {
         }else {
             return null;
         }
-//        Optional<Company> optionalCompany = companyRepository.findById(1L);
+    }
+
+    //회사 정보수정
+    @Transactional
+    public CompanyDto updateCompany(Company company, CompanyDto companyDto){
 //        optionalCompany.orElseThrow(()->new NotFoundException("not found company"));
-//
 //        if (optionalCompany.isPresent()){
-//            CompanyDto companyDto = new CompanyDto(optionalCompany.get());
-//            return companyDto;
+            company.update(companyDto);
+            CompanyDto companyDto1 = new CompanyDto(company);
+            return companyDto1;
 //        }else {
 //            return null;
 //        }
     }
 
-    //회사 정보수정
-    @Transactional
-    public CompanyDto updateCompany(Long companyId, CompanyDto companyDto){
-        Optional<Company> optionalCompany = companyRepository.findById(companyId);
-        optionalCompany.orElseThrow(()->new NotFoundException("not found company"));
-
-        if (optionalCompany.isPresent()){
-            optionalCompany.get().update(companyDto);
-            CompanyDto companyDto1 = new CompanyDto(optionalCompany.get());
-            return companyDto1;
-        }else {
-            return null;
-        }
-    }
-
     //회사 탈퇴
     @Transactional
-    public void deleteCompany(Long companyId){
-        Optional<Company> optionalCompany = companyRepository.findById(companyId);
-        optionalCompany.orElseThrow(()->new NotFoundException("not found company"));
-
-        companyRepository.deleteById(companyId);
+    public void deleteCompany(Company company){
+//        Optional<Company> optionalCompany = companyRepository.findById(company.getId());
+//        optionalCompany.orElseThrow(()->new NotFoundException("not found company"));
+        companyRepository.deleteById(company.getId());
     }
 
     //회사 공고 등록
     @Transactional
-    public JobOpeningDto postJobOpening(Long companyId, JobOpeningDto jobOpeningDto){
+    public JobOpeningDto postJobOpening(Company company, JobOpeningDto jobOpeningDto){
         Gugun gugun = gugunRepository.findById(jobOpeningDto.getGugunId()).get();
         JobChildCategory jobChildCategory = jobChildCategoryRepository.findById(jobOpeningDto.getJobChildCategoryId()).get();
 
         JobOpening jobOpening = new JobOpening(jobOpeningDto,gugun,jobChildCategory);
-        Optional<Company> company = companyRepository.findById(companyId);
-        if(company.isPresent()){
-            jobOpening.updateCompany(company.get());
+//        Optional<Company> company = companyRepository.findById(companyId);
+//        if(company.isPresent()){
+            jobOpening.updateCompany(company);
             jobOpeningRepository.save(jobOpening);
 
-            CompanyDto companyDto = new CompanyDto(company.get());
+            CompanyDto companyDto = new CompanyDto(company);
 
             JobOpeningDto jobOpeningDto1 = new JobOpeningDto(jobOpening,
                     companyDto,
@@ -128,19 +123,19 @@ public class CompanyService {
                     JobChildCategoryResponse.response(jobChildCategory));
 
             return jobOpeningDto1;
-        }else {
-            return null;
-        }
+//        }else {
+//            return null;
+//        }
     }
 
     //회사 공고 목록
     @Transactional
-    public List<JobOpeningDto> getJobOpeningList(Long companyId){
-        Optional<Company> optionalCompany = companyRepository.findById(companyId);
-        optionalCompany.orElseThrow(()->new NotFoundException("not found company"));
+    public List<JobOpeningDto> getJobOpeningList(Company company){
+//        Optional<Company> optionalCompany = companyRepository.findById(companyId);
+//        optionalCompany.orElseThrow(()->new NotFoundException("not found company"));
 
-        if (optionalCompany.isPresent()){
-            List<JobOpening> jobOpeningList = jobOpeningRepository.findByCompany(optionalCompany.get());
+//        if (optionalCompany.isPresent()){
+            List<JobOpening> jobOpeningList = jobOpeningRepository.findByCompany(company);
             List<JobOpeningDto> jobOpeningDtoList = jobOpeningList.stream().map(j->new JobOpeningDto(
                     j,
                     new CompanyDto(j.getCompany()),
@@ -149,15 +144,18 @@ public class CompanyService {
             )).collect(Collectors.toList());
 
             return jobOpeningDtoList;
-        }else {
-            return null;
-        }
+//        }else {
+//            return null;
+//        }
 
     }
 
     //회사 공고 상세조회
     @Transactional
-    public JobOpeningDto getJobOpening(Long jobOpeningId){
+    public JobOpeningDto getJobOpening(Company company, Long jobOpeningId){
+        if (company.getId() != jobOpeningRepository.findById(jobOpeningId).get().getCompany().getId())
+            throw new NotAuthException(COMPANY_NOT_MATCH);
+
         Optional<JobOpening> optionalJobOpening = jobOpeningRepository.findById(jobOpeningId);
         optionalJobOpening.orElseThrow(()->new NotFoundException("not found jobOpening"));
 
@@ -180,7 +178,10 @@ public class CompanyService {
 
     //회사 공고 수정
     @Transactional
-    public JobOpeningDto updateJobOpening(Long jobOpeningId, JobOpeningDto jobOpeningDto){
+    public JobOpeningDto updateJobOpening(Company company, Long jobOpeningId, JobOpeningDto jobOpeningDto){
+        if (company.getId() != jobOpeningRepository.findById(jobOpeningId).get().getCompany().getId())
+            throw new NotAuthException(COMPANY_NOT_MATCH);
+
         Gugun gugun = gugunRepository.findById(jobOpeningDto.getGugunId()).get();
         JobChildCategory jobChildCategory = jobChildCategoryRepository.findById(jobOpeningDto.getJobChildCategoryId()).get();
 
@@ -190,8 +191,8 @@ public class CompanyService {
         if (optionalJobOpening.isPresent()){
             optionalJobOpening.get().update(jobOpeningDto,gugun,jobChildCategory);
 
-            Optional<Company> optionalCompany = companyRepository.findById(1L);
-            CompanyDto companyDto = new CompanyDto(optionalCompany.get());
+//            Optional<Company> optionalCompany = companyRepository.findById(1L);
+            CompanyDto companyDto = new CompanyDto(company);
 
             JobOpeningDto jobOpeningDto1 = new JobOpeningDto(optionalJobOpening.get(),
                     companyDto,
@@ -206,17 +207,63 @@ public class CompanyService {
 
     //회사 공고 삭제
     @Transactional
-    public void deleteJobOpening(Long jobOpeningId){
+    public void deleteJobOpening(Company company, Long jobOpeningId){
+        if (company.getId() != jobOpeningRepository.findById(jobOpeningId).get().getCompany().getId())
+            throw new NotAuthException(COMPANY_NOT_MATCH);
 
         Optional<JobOpening> optionalJobOpening = jobOpeningRepository.findById(jobOpeningId);
         optionalJobOpening.orElseThrow(()->new NotFoundException("not found JobOpening"));
 
         jobOpeningRepository.deleteById(jobOpeningId);
     }
+
     //회사 공고 지원자 목록
+    @Transactional
+    public List<ApplyDto> getappliyList(Company company, Long jobOpeningId){
+        if (company.getId() != jobOpeningRepository.findById(jobOpeningId).get().getCompany().getId())
+            throw new NotAuthException(COMPANY_NOT_MATCH);
+
+        JobOpening jobOpening = jobOpeningRepository.findById(jobOpeningId).orElseThrow(() -> new NotFoundException("not found Apply"));
+        JobOpeningDto jobOpeningDto = new JobOpeningDto(
+                jobOpening,
+                new CompanyDto(company),
+                GugunResponse.response(jobOpening.getGugun()),
+                JobChildCategoryResponse.response(jobOpening.getJobChildCategory())
+                );
+
+        List<Apply> applyList = applyRepository.findByJobOpening(jobOpening);
+        List<ApplyDto> applyDtoList = applyList.stream().map(a->new ApplyDto(
+                a,
+                jobOpeningDto,
+                UserResponse.response(a.getUser())
+        )).collect(Collectors.toList());
+
+        return  applyDtoList;
+    }
+
     //회사 공고 지원자 상세목록
+    @Transactional
+    public UserResponse getapplicant(Company company, Long applyId){
+        if(company.getId() != applyRepository.findById(applyId).get().getJobOpening().getCompany().getId())
+            throw new NotAuthException(COMPANY_NOT_MATCH);
+
+        Apply apply = applyRepository.findById(applyId).orElseThrow(() -> new NotFoundException("not found Apply"));
+        UserResponse userResponse = UserResponse.response(apply.getUser());
+
+        return userResponse;
+    }
+
     //회사 공고 지원자 상태수정
+    @Transactional
+    public Apply updateapply(Company company, Long applyId, ApplyDto applyDto){
+        if(company.getId() != applyRepository.findById(applyId).get().getJobOpening().getCompany().getId())
+            throw new NotAuthException(COMPANY_NOT_MATCH);
 
+        Apply apply = applyRepository.findById(applyId).orElseThrow(() -> new NotFoundException("not found Apply"));
+        apply.update(applyDto,apply.getUser(), apply.getJobOpening());
 
+        return apply;
+
+    }
 
 }
